+++ src/main/java/sejong/coffee/yun/dto/pay/CardPaymentDto.java
@@ -1,33 +1,42 @@
 package sejong.coffee.yun.dto.pay;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
+import lombok.Getter;
 import sejong.coffee.yun.domain.order.Order;
 import sejong.coffee.yun.domain.pay.CardPayment;
 import sejong.coffee.yun.domain.pay.PaymentCancelReason;
 import sejong.coffee.yun.domain.pay.PaymentStatus;
 import sejong.coffee.yun.domain.user.Card;
+import sejong.coffee.yun.dto.order.OrderDto;
 import sejong.coffee.yun.infra.port.UuidHolder;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import java.time.LocalDateTime;
 
-import static sejong.coffee.yun.util.parse.ParsingDateTimeUtil.parsingCardValidDate;
+import static sejong.coffee.yun.util.parse.ParsingUtil.parsingCardValidDate;
+import static sejong.coffee.yun.util.parse.ParsingUtil.parsingMemberIdentityNumber;
 
 
 @Builder
+@Getter
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class CardPaymentDto {
+
     public record Request(
             @NotNull(message = "카드번호가 없습니다.")
             String cardNumber,
             @Pattern(regexp = "(?=.*[0-9]).{4}", message = "비밀번호는 숫자 4자여야 합니다.")
             String cardPassword,
-            @NotNull(message = "만료 연도가 없습니다.")
-            String cardExpirationYear,
             @NotNull(message = "만료 월이 없습니다.")
             String cardExpirationMonth,
+            @NotNull(message = "만료 연도가 없습니다.")
+            String cardExpirationYear,
             @NotNull(message = "주문 아이디가 없습니다.")
-            String orderUuid,
+            String orderId,
             @NotNull(message = "주문명이 없습니다.")
             String orderName,
             @NotNull(message = "주문 금액이 없습니다.")
@@ -35,13 +44,18 @@ public class CardPaymentDto {
             @NotNull(message = "카드 소유자 번호가없습니다.")
             String customerIdentityNumber,
             String customerName,
+            @JsonIgnore
             LocalDateTime requestedAt,
+            @JsonIgnore
             Order order
+            // 객체를 JSON 문자열로 변환
     ) {
         public static Request create(Card card, Order order, UuidHolder uuidHolder) {
-            return new CardPaymentDto.Request(card.getNumber(), card.getCardPassword(), parsingCardValidDate(card.getValidThru())[0],
-                    parsingCardValidDate(card.getValidThru())[1], uuidHolder.random(), order.getName(), order.getOrderPrice().getTotalPrice().toString(),
-                    card.getMember().getEmail(), card.getMember().getName(), LocalDateTime.now(), order);
+            return new CardPaymentDto.Request(card.getNumber(), card.getCardPassword().substring(0, 2), parsingCardValidDate(card.getValidThru())[0],
+                    parsingCardValidDate(card.getValidThru())[1], uuidHolder.random(), order.getName(), String.valueOf(order.getOrderPrice().getTotalPrice().intValue()),
+//                    parsingMemberIdentityNumber(card.getMember().getEmail()),
+                    parsingMemberIdentityNumber("231212"),
+                    card.getMember().getName(), LocalDateTime.now(), order);
         }
 
         public static Request from(CardPayment cardPayment) {
@@ -49,12 +63,14 @@ public class CardPaymentDto {
                     cardPayment.getCardExpirationYear(), cardPayment.getCardExpirationMonth(),
                     cardPayment.getOrderUuid(), cardPayment.getOrder().getName(),
                     cardPayment.getOrder().getOrderPrice().getTotalPrice().toString(),
-                    cardPayment.getOrder().getMember().getEmail().split("@")[0],
+//                    parsingMemberIdentityNumber(cardPayment.getOrder().getMember().getEmail()),
+                    parsingMemberIdentityNumber("231212"),
                     cardPayment.getOrder().getMember().getName(), LocalDateTime.now(), cardPayment.getOrder());
         }
     }
 
     public record Response(
+            @JsonProperty("orderId")
             String orderUuid,
             String orderName,
             String cardNumber,
@@ -62,10 +78,11 @@ public class CardPaymentDto {
             String cardExpirationMonth,
             String totalAmount,
             String paymentKey,
+//            @JsonIgnore
             PaymentStatus paymentStatus,
-            LocalDateTime requestedAt,
-            LocalDateTime approvedAt,
-            Order order,
+            String requestedAt,
+            String approvedAt,
+            OrderDto.Response orderDto,
             PaymentCancelReason cancelReason
     ) {
         public Response(CardPayment entity) {
@@ -73,14 +90,14 @@ public class CardPaymentDto {
                     entity.getCardNumber(), entity.getCardExpirationYear(), entity.getCardExpirationMonth(),
                     entity.getOrder().getOrderPrice().getTotalPrice().toString(),
                     entity.getPaymentKey(), entity.getPaymentStatus(),
-                    entity.getRequestedAt(), entity.getApprovedAt(), entity.getOrder(), null);
+                    entity.getRequestedAt().toString(), entity.getApprovedAt().toString(), new OrderDto.Response(entity.getOrder()), null);
         }
 
         public static Response cancel(CardPayment entity) {
             return new Response(entity.getOrderUuid(), entity.getOrder().getName(), entity.getCardNumber(),
                     entity.getCardExpirationYear(), entity.getCardExpirationMonth(),
                     entity.getOrder().getOrderPrice().getTotalPrice().toString(), entity.getPaymentKey(), entity.getPaymentStatus(),
-                    entity.getRequestedAt(), entity.getApprovedAt(), entity.getOrder(), entity.getCancelReason());
+                    entity.getRequestedAt().toString(), entity.getApprovedAt().toString(), new OrderDto.Response(entity.getOrder()), entity.getCancelReason());
         }
     }
 }
